Moved python files to 'bin' and added folders for root and gltf files to be added to. Tidied up the code a bit to make it more usable and readable
