--- edit_Ecal.json
@@ -1,172 +1,0 @@
-{
-    "childrenToHide": [],
-    "subParts": {
-        "BeamCal": [
-            [
-                "(BeamCal)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "Coil": [
-            [
-                "(Coil)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "EcalBarrel": [
-            [
-                "EcalBarrel_module\\w+"
-            ],
-            0.8],
-
-        "EcalEndcap": [
-            [
-                "EcalEndcapQuadrant\\w+"
-            ],
-            0.8
-        ],
-        "EcalPlug": [
-            [
-                "(EcalPlug)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "FTD": [
-            [
-                "(FTD)_(?!envelope)\\w+",
-                "FTDDisk_.*_negZ",
-                "ftd_petal_negZ_.*_.*",
-                "ftd_sensor_negZ_.*_.*_.*",
-                "FTDDisk_.*_posZ",
-                "ftd_petal_posZ_.*_.*",
-                "ftd_sensor_posZ_.*_.*_.*",
-                "FTD_support"
-            ],
-            0.8
-        ],
-        "HcalBarrel": [
-            [
-                "(HcalBarrel)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "HcalEndcap": [
-            [
-                "(HcalEndcap)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "HcalRing": [
-            [
-                "(HcalRing)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "LHCal": [
-            [
-                "(LHCal)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "Lcal": [
-            [
-                "(Lcal)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "QD0_cryostat": [
-            [
-                "(QD0_cryostat)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "QD0_support": [
-            [
-                "(QD0_support)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "SET": [
-            [
-                "(SET)_(?!envelope)\\w+",
-                "set_ladder_.*_.*",
-                "set_ladder_.*_.*_.*"
-            ],
-            0.8
-        ],
-        "SIT": [
-            [
-                "(SIT)_(?!envelope)\\w+",
-                "sit_ladder_.*_.*",
-                "sit_ladder_.*_.*_.*"
-            ],
-            0.8
-        ],
-        "SServices00": [
-            [
-                "(SServices00)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "TPC": [
-            [
-                "(TPC)_(?!envelope)\\w+",
-                "tpc_endcap_bwd",
-                "tpc_endcap_fwd",
-                "tpc_sensGas_bwd",
-                "tpc_row_bwd_.*",
-                "tpc_sensGas_fwd",
-                "tpc_row_fwd_.*"
-            ],
-            0.8
-        ],
-        "Tube": [
-            [
-                "(Tube)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "VXD": [
-            [
-                "(VXD)_(?!envelope)\\w+",
-                "BerylliumAnnulusBlock_.*_negZ_.*",
-                "BerylliumAnnulusBlock_.*_posZ_.*",
-                "ElectronicsEnd_.*_negZ_.*",
-                "ElectronicsEnd_.*_posZ_.*",
-                "SiActiveLayer_.*_negZ_.*",
-                "SiActiveLayer_.*_posZ_.*",
-                "VXD_support",
-                "KaptonLines_.*_negZ",
-                "KaptonLines_.*_posZ",
-                "MetalLines_.*_negZ",
-                "MetalLines_.*_posZ",
-                "VXD_ExtKaptonCab_neg",
-                "VXD_ExtKaptonCab_pos",
-                "VXD_ExtMetalCab_neg",
-                "VXD_ExtMetalCab_pos",
-                "VXD_endplate_bwd",
-                "VXD_endplate_fwd",
-                "VXD_support_bwd_inner",
-                "VXD_support_bwd_outer",
-                "VXD_support_fwd_inner",
-                "VXD_support_fwd_outer",
-                "negShellCone",
-                "posShellCone"
-            ],
-            0.8
-        ],
-        "YokeBarrel": [
-            [
-                "(YokeBarrel)_(?!envelope)\\w+"
-            ],
-            0.8
-        ],
-        "YokeEndcap": [
-            [
-                "(YokeEndcap)_(?!envelope)\\w+"
-            ],
-            0.8
-        ]
-    },
-    "maxLevel": 3
-}
